fixed rank requests

+++ server/getRanks.js
@@ -60,10 +60,10 @@ function returnAPIDATA() {
         request(optionsOPRS, function(error, response) {
             if (error) throw new Error(error)
             printMessage("Status Code", response.statusCode)
-            consoleLog(response.body)
+            //consoleLog(response.body)
             const oprData = JSON.parse(response.body)
             
-            consoleLog(oprData)
+            //consoleLog(oprData)
 
             for (const [rankings, _] of Object.entries(oprData)) {
                 for (const [i, val] of Object.entries(oprData[rankings])) {
@@ -74,31 +74,33 @@ function returnAPIDATA() {
             }
 
             request(optionsRankings, function(error, response) {
-                consoleLog(response.body)
                 const rankingsData = JSON.parse(response.body).rankings
                 //consoleLog(rankingsData)
                 const combinedTeamData = {}
+                if (rankingsData) {
 
-                for (let i = 0; i<rankingsData.length; i++) {
-                    combinedTeamData[rankingsData[i].team_key.substring(3)] = rankingsData[i]
-                    combinedTeamData[rankingsData[i].team_key.substring(3)].opr = oprData?.["oprs"]?.[rankingsData[i].team_key.substring(3)]
-                    combinedTeamData[rankingsData[i].team_key.substring(3)].dpr = oprData?.["dprs"]?.[rankingsData[i].team_key.substring(3)]
+                    for (let i = 0; i<rankingsData.length; i++) {
+                        combinedTeamData[rankingsData[i].team_key.substring(3)] = rankingsData[i]
+                        combinedTeamData[rankingsData[i].team_key.substring(3)].opr = oprData?.["oprs"]?.[rankingsData[i].team_key.substring(3)]
+                        combinedTeamData[rankingsData[i].team_key.substring(3)].dpr = oprData?.["dprs"]?.[rankingsData[i].team_key.substring(3)]
+                    }
+
+                    //consoleLog(database.writeAPIData(combinedTeamData))
+                    //consoleLog(combinedTeamData)    
+                    if (Object.keys(combinedTeamData).length) { 
+                        database.query(database.deleteAPIData(), (err, res) => {
+                            consoleLog(err)
+                            //consoleLog(res)
+                            database.query(database.writeAPIData(combinedTeamData), (err, res) => {
+                                consoleLog(err)
+                                //consoleLog(res)
+                            })
+                        })
+                    }
+        
+                    
+                    //consoleLog(combinedTeamData)
                 }
-
-                //consoleLog(database.writeAPIData(combinedTeamData))
-                //consoleLog(combinedTeamData)    
-                
-                database.query(database.deleteAPIData(), (err, res) => {
-                    consoleLog(err)
-                    //consoleLog(res)
-                    database.query(database.writeAPIData(combinedTeamData), (err, res) => {
-                        consoleLog(err)
-                        //consoleLog(res)
-                    })
-                })
-    
-                
-                //consoleLog(combinedTeamData)
                 resolve(combinedTeamData)
             })
